Properties take two
Finish implementation of property parsing and XML generation.

With this implementation, there are two property maps, one in the base
class and the other in the derived concrete item. Derived classes use
two functions to access the properties in the base class:

 - item::parse_item_properties and
 - item::item_properties_to_xml

For example:

  +-------+
  | item  |
  +---+---+  properties_ hash map
      ^
      |      protected:
      |        void parse_item_properties(const rapidxml::xml_node<>&)
      |        std::string item_properties_to_xml() const
  +---+---+
  | task  |
  +-------+   properties_ hash map

Still a lot of code duplication with this commit.

Also, a task's <Owner/> is a read-only property and is somehow not in
returned payload when testing against Exchange 2013.