Implement notification operations (#85)
This commit implements EWS notification operations as in

    https://msdn.microsoft.com/en-us/library/office/bb409286(v=exchg.150).aspx#Anchor_9

Following operations are done:
- <Subscribe>
- <Unsubscribe>
- <GetEvents>

Not included in this commit:
- <GetStreamingEvents>

The event_type enum and internal::base_event class-hierarchy has been added to
represent all events. The subscription_information struct contains the
subscription ID and watermark which are needed to further query the Exchange
server for updates. This object is returned by the basic_service::subscribe()
member-function. Tests have been added as well to check the validity of the
parsed information and the generated XML.

Code uses C++17 features heavily, mostly std::variant. So it is only
compiled-in when EWS_HAS_VARIANT is set during compilation. There is also a
feature-test macro that tests for std::variant-support and sets this
preprocessor macro when available.

Fixes issue #80.