Add use of curl error buffer
curl_ptr now registers a buffer for curl for detailed error
information. This will be forwarded into the curl_error in case of an
error when calling curl_easy_perform. curl_error, in turn, is moved out
of the internal namespace to allow the user to catch this specific
exception.