replace nlohmann/json with Tencent/rapidjson, fix oauth scope vs. resource
as nlohman/json seems to have problems with the current
os/compiler combinations, i replaced it with
https://github.com/Tencent/rapidjson

the final inclusion of the additional libraries can happen
via some package management as discussed previously
in this thread, for now i'll just copy it like before.

this fixes also that scope was being used instead of resource,
the scope parameter isn't used for client credentials flow.