Add option to turn off building tests and examples.
- Add option to disable/enable generating tests (Default is on).
- Add option to disable/enable generating examples (Default is on).

The option GENERATE_TESTS is quite straightforward, when ON tests are
built, when OFF tests aren't build.

The option GENERATE_EXAMPLES is rather different, when ON examples are
able to be build, by default all examples are build when this is
ON. Option GENERATE_ALL_EXAMPLES disables that, when
GENERATE_ALL_EXAMPLES is ON all examples are built, but only done so when
GENERATE_EXAMPLES is also ON.

To build specific examples, e.g: 'X', GENERATE_EXAMPLES must be on, option
GENERATE_EXAMPLE_x is ON, and OFF if there is no need to build 'X'
example.

To add more examples, simply add
option(GENERATE_EXAMPLES_new_example_name "Example description"
${GENERATE_EXAMPLES}) to it's appropiate place (In example's triggers,
chunk). Add: add_example(new_example_name) to it's normal place (In the
list of examples). And
then the new option can be enabled/disabled as described above.